Poser le filet : rien ne part plus sans avoir été testé
Les 132 tests du dépôt existaient, ils étaient bons, et ils ne tournaient que
quand quelqu'un y pensait : ni release.yml ni docker.yml n'exécutait cargo
test ou cargo clippy. On publiait un binaire compilé, jamais un binaire
testé — et une régression dans le tampon de gigue, dans les rôles ou dans le
rééchantillonneur pouvait partir en release sans qu'aucun signal ne se lève.

ci.yml vérifie donc l'espace de travail complet sur Windows — le seul système
où tout compile, WASAPI et DPAPI étant derrière un #[cfg(windows)] — et le
serveur seul sur Linux, là où il tourne pour de vrai : c'est cette passe qui
attrape un chemin écrit à la Windows ou un #[cfg] posé à l'envers.

release.yml ET docker.yml en dépendent. L'image du serveur est gatée elle
aussi, et c'est le cas le plus pressant : elle se déploie toute seule
(Watchtower sonde toutes les cinq minutes), donc une régression y serait en
production avant que quiconque l'ait constatée. Elle appelle ci.yml avec
only_server pour ne pas attendre derrière openh264 et une interface graphique
dont elle ne dépend pas.

La clé de concurrence porte github.workflow, qui dans un workflow appelé
nomme l'APPELANT : sur une poussée de tag, release.yml et docker.yml
déclenchent tous les deux ci.yml avec le même github.ref, et sans cette
moitié de clé ils s'annulaient mutuellement. On n'annule pas sur un tag non
plus — une release interrompue ne se rattrape pas toute seule.

Les dix-sept avertissements de clippy sont soldés pour que -D warnings puisse
exister. require_admin était bien mort ; exists ne l'était pas — clippy le
voyait mort dans la passe non-test, mais un test l'utilisait, et il dit
désormais la même chose avec get(), qui est de l'API vivante.

Une correction dépasse le lint : roster() trie par sort_by_cached_key et non
sort_by_key. La clé est une String, donc une allocation, et sort_by_key
l'aurait recalculée à chaque comparaison — sur une liste rediffusée à tout le
serveur à chaque entrée et sortie de vocal.

cargo fmt --check reste volontairement hors du filet : la base n'est pas au
format rustfmt (419 hunks d'écart, 242 avec la meilleure configuration
trouvée), et son style manuel est délibéré — il garde les motifs de match sur
une ligne là où rustfmt les éclate en six. Adopter rustfmt reste possible,
mais c'est une décision à prendre pour elle-même, dans son propre commit avec
un .git-blame-ignore-revs, pas un effet de bord de l'introduction de la CI.

+++ crates/voice/src/wasapi.rs
@@ -328,7 +328,7 @@ unsafe fn parse_format(fmt: *const WAVEFORMATEX) -> anyhow::Result<(SampleKind, 
     let channels = (*fmt).nChannels;
     let kind = if tag == WAVE_FORMAT_IEEE_FLOAT && bits == 32 {
         SampleKind::F32
-    } else if tag == WAVE_FORMAT_PCM as u32 && bits == 16 {
+    } else if tag == WAVE_FORMAT_PCM && bits == 16 {
         SampleKind::I16
     } else if tag == WAVE_FORMAT_EXTENSIBLE {
         let ext = fmt as *const WAVEFORMATEXTENSIBLE;
